fix(decomp3D):Fixed issues with removing edges for the decomp3D class, disabled multithreading due to overheads and output safety.

+++ src/DecompHeaders/Block.h
@@ -5,7 +5,7 @@ class Block
 {
 public:
     int zDimension, yDimension, xDimension;
-    int zStart, yStart, xStart;
+    size_t zStart, yStart, xStart;
 };
 
 #endif
